+++ api/Data/Entity/Contrived/Budget.cs
@@ -6,5 +6,5 @@ public class Budget
     public int TeamId { get; set; }
 
     // a new week begins at midnight of the first game of that week (typically Thursday). 
-    public required decimal[] PaymentSchedule { get; set; }
+    public required float[] PaymentSchedule { get; set; }
 }
