+++ api/Data/Models/StatSet.cs
@@ -6,7 +6,7 @@ public record StatSet
 {
     public required StatSetKind Kind { get; init; }
     public required List<StatScore> Scores { get; init; }
-    public decimal Total => Scores.Sum(s => s.Value);
+    public float Total => Scores.Sum(s => s.Value);
 
     public static StatSet From(StatSetKind kind, List<StatScore> scores) => new()
     {
