+++ api/Infrastructure/Interfaces/IBudgetRepository.cs
@@ -6,5 +6,5 @@ public interface IBudgetRepository
 {
     Task<Budget?> GetByTeamIdAsync(int teamId, CancellationToken cancellationToken = default);
     Task<List<Budget>> GetByLeagueIdAsync(int leagueId, CancellationToken cancellationToken = default);
-    Task<bool> UpdateAsync(int teamId, decimal[] paymentSchedule, CancellationToken cancellationToken = default);
+    Task<bool> UpdateAsync(int teamId, float[] paymentSchedule, CancellationToken cancellationToken = default);
 }
